implement email verifiaction and clean hardcoding
