+++ lib/constants/routes.dart
@@ -1,0 +1,4 @@
+const registerRoute = '/register/';
+const loginRoute = '/login/';
+const notesRoute = '/notes/';
+const verifyEmailRoute = '/verifyEmail/';
